changed from Double to BigDecimal

+++ src/main/kotlin/io/github/frantoso/physicalquantities/core/Prefix.kt
@@ -1,29 +1,31 @@
 package io.github.frantoso.physicalquantities.core
 
+import java.math.BigDecimal
+
 // @formatter:off
 @Suppress("ktlint")
 enum class Prefix(
     val symbol: String,
-    val factorToBase: Double,
-    val factorFromBase: Double,
+    val factorToBase: BigDecimal,
+    val factorFromBase: BigDecimal,
 ) {
-    ATTO("a", 1e-18, 1e18),
-    FEMTO("f", 1e-15, 1e15),
-    PICO("p", 1e-12, 1e12),
-    NANO("n", 1e-9, 1e9),
-    MICRO("µ", 1e-6, 1e6),
-    MILLI("m", 1e-3, 1e3),
-    CENTI("c", 1e-2, 1e2),
-    DECI("d", 0.1, 10.0),
-    DECA("da", 10.0, 0.1),
-    HECTO("h", 1e2, 1e-2),
-    KILO("k", 1e3, 1e-3),
-    MEGA("M", 1e6, 1e-6),
-    GIGA("G", 1e9, 1e-9),
-    TERA("T", 1e12, 1e-12),
-    PETA("P", 1e15, 1e-15),
-    EXA("E", 1e18, 1e-18),
-    None("", 1.0, 1.0),
+    ATTO("a", 1e-18.toBigDecimal(), 1e18.toBigDecimal()),
+    FEMTO("f", 1e-15.toBigDecimal(), 1e15.toBigDecimal()),
+    PICO("p", 1e-12.toBigDecimal(), 1e12.toBigDecimal()),
+    NANO("n", 1e-9.toBigDecimal(), 1e9.toBigDecimal()),
+    MICRO("µ", 1e-6.toBigDecimal(), 1e6.toBigDecimal()),
+    MILLI("m", 1e-3.toBigDecimal(), 1e3.toBigDecimal()),
+    CENTI("c", 1e-2.toBigDecimal(), 1e2.toBigDecimal()),
+    DECI("d", 0.1.toBigDecimal(), 10.0.toBigDecimal()),
+    DECA("da", 10.0.toBigDecimal(), 0.1.toBigDecimal()),
+    HECTO("h", 1e2.toBigDecimal(), 1e-2.toBigDecimal()),
+    KILO("k", 1e3.toBigDecimal(), 1e-3.toBigDecimal()),
+    MEGA("M", 1e6.toBigDecimal(), 1e-6.toBigDecimal()),
+    GIGA("G", 1e9.toBigDecimal(), 1e-9.toBigDecimal()),
+    TERA("T", 1e12.toBigDecimal(), 1e-12.toBigDecimal()),
+    PETA("P", 1e15.toBigDecimal(), 1e-15.toBigDecimal()),
+    EXA("E", 1e18.toBigDecimal(), 1e-18.toBigDecimal()),
+    None("", 1.0.toBigDecimal(), 1.0.toBigDecimal()),
 }
 // @formatter:on
 
@@ -32,5 +34,6 @@ enum class Prefix(
  * @throws NoSuchPrefixException in case of an unknown prefix string.
  */
 fun String.toPrefix(): Prefix =
-    enumValues<Prefix>().firstOrNull { it.name == this.uppercase() } ?: enumValues<Prefix>().firstOrNull { it.symbol == this }
+    enumValues<Prefix>().firstOrNull { it.name == this.uppercase() }
+        ?: enumValues<Prefix>().firstOrNull { it.symbol == this }
         ?: throw NoSuchPrefixException("Unknown prefix $this")
