+++ src/test/kotlin/io/github/frantoso/physicalquantities/electrical/PowerTest.kt
@@ -2,6 +2,7 @@ package io.github.frantoso.physicalquantities.electrical
 
 import io.github.frantoso.physicalquantities.core.ValueWithUnit
 import io.github.frantoso.physicalquantities.core._m
+import io.github.frantoso.physicalquantities.utils.toBigDecimal
 import org.assertj.core.api.Assertions.assertThat
 import kotlin.test.Test
 
@@ -10,21 +11,21 @@ class PowerTest {
     fun `test initialisation`() {
         val power = Power.fromWatt(24)
 
-        assertThat(power.watt).isEqualTo(24.0)
+        assertThat(power.watt).isEqualTo(24.toBigDecimal())
     }
 
     @Test
     fun `test initialisation from literal`() {
         val power = 24.W
 
-        assertThat(power.watt).isEqualTo(24.0)
+        assertThat(power.watt).isEqualTo(24.toBigDecimal())
     }
 
     @Test
     fun `test createFromValue`() {
         val power = -(24.W) // base class calls createFromValue()
 
-        assertThat(power.watt).isEqualTo(-24.0)
+        assertThat(power.watt).isEqualTo((-24).toBigDecimal())
     }
 
     @Test
