+++ src/test/kotlin/io/github/frantoso/physicalquantities/electrical/VoltageTest.kt
@@ -2,6 +2,7 @@ package io.github.frantoso.physicalquantities.electrical
 
 import io.github.frantoso.physicalquantities.core.ValueWithUnit
 import io.github.frantoso.physicalquantities.core._m
+import io.github.frantoso.physicalquantities.utils.toBigDecimal
 import org.assertj.core.api.Assertions.assertThat
 import kotlin.test.Test
 
@@ -10,21 +11,21 @@ class VoltageTest {
     fun `test initialisation`() {
         val voltage = Voltage.fromVolt(24)
 
-        assertThat(voltage.volt).isEqualTo(24.0)
+        assertThat(voltage.volt).isEqualTo(24.toBigDecimal())
     }
 
     @Test
     fun `test initialisation from literal`() {
         val voltage = 24.V
 
-        assertThat(voltage.volt).isEqualTo(24.0)
+        assertThat(voltage.volt).isEqualTo(24.toBigDecimal())
     }
 
     @Test
     fun `test createFromValue`() {
         val voltage = -(24.V) // base class calls createFromValue()
 
-        assertThat(voltage.volt).isEqualTo(-24.0)
+        assertThat(voltage.volt).isEqualTo((-24).toBigDecimal())
     }
 
     @Test
